Move clean function to helpers file
Avoids a circular import when importing api.facility_type_processing_type
variables into api.models.

Refs #1590
