+++ src/django/api/helpers.py
@@ -1,6 +1,7 @@
 import re
 import csv
 import json
+from unidecode import unidecode
 
 CONSONANT_SOUND = re.compile(r'''
 one(![ir])
@@ -61,3 +62,23 @@ def get_list_contributor_field_values(item, fields):
             index = list_fields.index(f['column_name'])
             f['value'] = data_values[index]
     return fields
+
+
+def clean(column):
+    """
+    Remove punctuation and excess whitespace from a value before using it to
+    find matches. This should be the same function used when developing the
+    training data read from training.json as part of train_gazetteer.
+    """
+    column = unidecode(column)
+    column = re.sub('\n', ' ', column)
+    column = re.sub('-', '', column)
+    column = re.sub('/', ' ', column)
+    column = re.sub("'", '', column)
+    column = re.sub(",", '', column)
+    column = re.sub(":", ' ', column)
+    column = re.sub(' +', ' ', column)
+    column = column.strip().strip('"').strip("'").lower().strip()
+    if not column:
+        column = None
+    return column
